src/requirements.txt
aws-cdk-lib==2.204.0
aws-cdk.aws-lambda-python-alpha==2.204.0a0
boto3==1.39.3
constructs==10.4.2
requests==2.32.4
